Fix migrations
And change percentage to f64 float

Source: https://github.com/msupply-foundation/open-msupply/commit/49e3badf92777bb1ce97dfbf63d13e22d9fce9fa

+++ server/graphql/demographic/src/mutations/insert_indicator.rs
@@ -51,7 +51,7 @@ pub struct InsertDemographicIndicatorInput {
     pub name: String,
     pub base_year: i32,
     pub base_population: Option<i32>,
-    pub population_percentage: Option<f32>,
+    pub population_percentage: Option<f64>,
     pub year_1_projection: Option<i32>,
     pub year_2_projection: Option<i32>,
     pub year_3_projection: Option<i32>,
