+++ server/service/src/demographic/update_demographic_indicator.rs
@@ -20,7 +20,7 @@ pub struct UpdateDemographicIndicator {
     pub name: Option<String>,
     pub base_year: Option<i32>,
     pub base_population: Option<i32>,
-    pub population_percentage: Option<f32>,
+    pub population_percentage: Option<f64>,
     pub year_1_projection: Option<i32>,
     pub year_2_projection: Option<i32>,
     pub year_3_projection: Option<i32>,
